upd hasValidReturnTypeKDoc()
