+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/provider/KoReturnTypeProviderCore.kt
@@ -1,6 +1,7 @@
 package com.lemonappdev.konsist.core.provider
 
 import com.lemonappdev.konsist.api.declaration.KoTypeDeclaration
+import com.lemonappdev.konsist.api.provider.KoKDocProvider
 import com.lemonappdev.konsist.api.provider.KoReturnTypeProvider
 import com.lemonappdev.konsist.core.util.ReceiverUtil
 import org.jetbrains.kotlin.psi.KtFunction
@@ -10,6 +11,7 @@ import org.jetbrains.kotlin.psi.psiUtil.isExtensionDeclaration
 internal interface KoReturnTypeProviderCore :
     KoReturnTypeProvider,
     KoContainingDeclarationProviderCore,
+    KoKDocProviderCore,
     KoBaseProviderCore {
     val ktFunction: KtFunction
